# Refactor chapel-py to use the C Python Stable API