--- a/tools/chapel-py/src/core-types.cpp
+++ b/tools/chapel-py/src/core-types.cpp
@@ -26,6 +26,7 @@
 #include "python-types.h"
 #include "error-tracker.h"
 #include "resolution.h"
+#include "python-type-helper.h"
 
 using namespace chpl;
 using namespace uast;
@@ -105,7 +106,7 @@ std::string generatePyiFile() {
     printedAnything = false; \
     generated.insert(NODE##Object::Name); \
     if (auto parentType = ParentTypeInfo<NODE##Object>::parentTypeObject()) { \
-      ss << "(" << parentType->tp_name << ")"; \
+      ss << "(" /*TODO:<< parentType->tp_name*/ << ")"; \
     } \
     ss << ":" << std::endl;
   #define METHOD(NODE, NAME, DOCSTR, TYPEFN, BODY) \
@@ -132,7 +133,8 @@ std::string generatePyiFile() {
     if(generated.find(NODE##Object::Name) == generated.end()) { \
       ss << "class " << NODE##Object::Name; \
       if (auto parentType = ParentTypeInfo<NODE##Object>::parentTypeObject()) { \
-        ss << "(" << parentTypeFor(TAG)->tp_name << ")"; \
+        /*auto tp_name = PyType_GetSlot(parentTypeFor(TAG), Py_tp_name);*/ \
+        ss << "(" /*TODO<< tp_name*/ << ")"; \
       } \
       ss << ":" << std::endl; \
       ss << "    pass" << std::endl; \
@@ -155,7 +157,7 @@ PyObject* AstNodeObject::iter(AstNodeObject *self) {
 
 void ChapelTypeObject_dealloc(ChapelTypeObject* self) {
   Py_XDECREF(self->contextObject);
-  Py_TYPE(self)->tp_free((PyObject *) self);
+  call_tp_free(ChapelTypeObject::PythonType, (PyObject*) self);
 }
 
 PyObject* ChapelTypeObject::str(ChapelTypeObject* self) {
@@ -186,14 +188,14 @@ PyTypeObject* parentTypeFor(asttags::AstTag tag) {
 #define AST_BEGIN_SUBCLASSES(NAME)
 #define AST_END_SUBCLASSES(NAME) \
   if (tag > asttags::START_##NAME && tag < asttags::END_##NAME) { \
-    return &NAME##Type; \
+    return NAME##Type; \
   }
 #include "chpl/uast/uast-classes-list.h"
 #undef AST_NODE
 #undef AST_LEAF
 #undef AST_BEGIN_SUBCLASSES
 #undef AST_END_SUBCLASSES
-  return &AstNodeObject::PythonType;
+  return AstNodeObject::PythonType;
 }
 
 PyTypeObject* parentTypeFor(types::typetags::TypeTag tag) {
@@ -202,18 +204,18 @@ PyTypeObject* parentTypeFor(types::typetags::TypeTag tag) {
 #define TYPE_BEGIN_SUBCLASSES(NAME)
 #define TYPE_END_SUBCLASSES(NAME) \
   if (tag > types::typetags::START_##NAME && tag < types::typetags::END_##NAME) { \
-    return &NAME##Type; \
+    return NAME##Type; \
   }
 #include "chpl/types/type-classes-list.h"
 #undef TYPE_NODE
 #undef BUILTIN_TYPE_NODE
 #undef TYPE_BEGIN_SUBCLASSES
 #undef TYPE_END_SUBCLASSES
-  return &ChapelTypeObject::PythonType;
+  return ChapelTypeObject::PythonType;
 }
 
 PyTypeObject* parentTypeFor(chpl::types::paramtags::ParamTag tag) {
-  return &ParamObject::PythonType;
+  return ParamObject::PythonType;
 }
 
 PyObject* wrapGeneratedType(ContextObject* context, const AstNode* node) {
@@ -226,7 +228,7 @@ PyObject* wrapGeneratedType(ContextObject* context, const AstNode* node) {
   switch (node->tag()) {
 #define CAST_TO(NAME) \
     case asttags::NAME: \
-      toReturn = PyObject_CallObject((PyObject*) &NAME##Type, args); \
+      toReturn = PyObject_CallObject((PyObject*) NAME##Type, args); \
       ((NAME##Object*) toReturn)->parent.value_ = node->to##NAME(); \
       break;
 #define AST_NODE(NAME) CAST_TO(NAME)
@@ -256,7 +258,7 @@ PyObject* wrapGeneratedType(ContextObject* context, const types::Type* node) {
   switch (node->tag()) {
 #define CAST_TO(NAME) \
     case types::typetags::NAME: \
-      toReturn = PyObject_CallObject((PyObject*) &NAME##Type, args); \
+      toReturn = PyObject_CallObject((PyObject*) NAME##Type, args); \
       ((NAME##Object*) toReturn)->parent.value_ = (const types::Type*) node->to##NAME(); \
       break;
 #define TYPE_NODE(NAME) CAST_TO(NAME)
@@ -286,7 +288,7 @@ PyObject* wrapGeneratedType(ContextObject* context, const chpl::types::Param* no
   switch (node->tag()) {
 #define PARAM_NODE(NAME, TYPE) \
     case chpl::types::paramtags::NAME: \
-      toReturn = PyObject_CallObject((PyObject*) &NAME##Type, args); \
+      toReturn = PyObject_CallObject((PyObject*) NAME##Type, args); \
       ((NAME##Object*) toReturn)->parent.value_ = node->to##NAME(); \
       break;
 #include "chpl/types/param-classes-list.h"