--- a/tools/chapel-py/src/iterator-support.h
+++ b/tools/chapel-py/src/iterator-support.h
@@ -20,8 +20,7 @@
 #ifndef CHAPEL_PY_ITERATOR_SUPPORT_H
 #define CHAPEL_PY_ITERATOR_SUPPORT_H
 
-#define PY_SSIZE_T_CLEAN
-#include "Python.h"
+#include "PythonWrapper.h"
 #include "chpl/framework/Context.h"
 #include "core-types.h"